Custom Countdown
A form that allows you to give a name and choose a date up to which the app can count how much time is needed to reach.
Additional functionalities include the fact that you can save one event to localStorage and even after returning to the site
the specific countdown will be available.
--Int ticket 2 fixed an issue with the favicon

[Ticket: 2]